requirements.txt
-i https://pypi.org/simple
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
automat==20.2.0
cachetools==4.1.1; python_version ~= '3.5'
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.2.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cssselect==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
flask-cors==3.0.8
flask-jwt-extended==3.24.1
flask-login==0.5.0
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
flask==1.1.2
google-api-core==1.23.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
google-api-python-client==1.12.5
google-auth-httplib2==0.0.4
google-auth==1.23.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
googleapis-common-protos==1.52.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
gunicorn==20.0.4
httplib2==0.18.1
hyperlink==20.0.1
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
incremental==17.5.0
itemadapter==0.2.0; python_version >= '3.6'
itemloaders==1.0.4; python_version >= '3.6'
itsdangerous==1.1.0
jinja2==2.11.2
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
lxml==4.6.1; platform_python_implementation == 'CPython'
markupsafe==1.1.1
musicbrainzngs==0.7.1
newsapi-python==0.2.6
parsel==1.6.0
protego==0.1.16; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
protobuf==3.14.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pydispatcher==2.0.5
pyhamcrest==2.0.2; python_version >= '3.5'
pyjwt==1.7.1
pyopenssl==19.1.0
python-dotenv==0.14.0
pytz==2020.4
queuelib==1.5.0
requests==2.24.0
rsa==4.6; python_version >= '3.5'
scrapy==2.4.0
service-identity==18.1.0
six==1.15.0
sqlalchemy==1.3.19
twisted==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.25.11; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
w3lib==1.22.0
werkzeug==1.0.1
wtforms==2.3.3
zope.interface==5.2.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
